# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SAHU CSC FV1 - CSC Business Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:

    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated

  # Profile
...
  /ledger:
    get:
      operationId: listLedgerEntries
      tags: [ledger]
      summary: List ledger entries
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: serviceType
          in: query
          schema:
            type: string
        - name: customerName
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of ledger entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerListResponse"
    post:
      operationId: createLedgerEntry
      tags: [ledger]
      summary: Create ledger entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LedgerEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerEntry"
...
  /ledger/summary:
    get:
      operationId: getLedgerSummary
      tags: [ledger]
      summary: Get ledger summary stats
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, yesterday, week, month, custom]
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerSummary"
  /ledger/{id}:
    get:
      operationId: getLedgerEntry
      tags: [ledger]
      summary: Get ledger entry by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ledger entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerEntry"
        "404":
          description: Not found
    patch:
      operationId: updateLedgerEntry
      tags: [ledger]
      summary: Update ledger entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LedgerEntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerEntry"
        "404":
          description: Not found
    delete:
      operationId: deleteLedgerEntry
      tags: [ledger]
      summary: Delete ledger entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # Services
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: Not found
  /notifications/{id}:
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete notification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # Audit Logs
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Username, email or mobile
        password:
          type: string
    AuthUser:
      type: object
      required: [id, username, email, role]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        mobile:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, operator, user]
        fullName:
          type: ["string", "null"]
        profilePicture:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
...
    LedgerEntry:
      type: object
      required: [id, date, customerName, serviceType, credit, debit, description, balance, createdBy, createdAt]
      properties:
        id:
          type: integer
        date:
          type: string
        customerName:
          type: string
        serviceType:
          type: string
        credit:
          type: number
        debit:
          type: number
        description:
          type: string
        balance:
          type: number
        createdBy:
          type: integer
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
    LedgerEntryInput:
      type: object
      required: [date, customerName, serviceType, credit, debit, description]
      properties:
        date:
          type: string
        customerName:
          type: string
        serviceType:
          type: string
        credit:
          type: number
        debit:
          type: number
        description:
          type: string
    LedgerEntryUpdate:
      type: object
      properties:
        date:
          type: string
        customerName:
          type: string
        serviceType:
          type: string
        credit:
          type: number
        debit:
          type: number
        description:
          type: string
    LedgerListResponse:
      type: object
      required: [entries, total, page, limit]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LedgerEntry"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    LedgerSummary:
      type: object
      required: [totalTransactions, totalCredits, totalDebits, netChange, period]
      properties:
        totalTransactions:
          type: integer
        totalCredits:
          type: number
        totalDebits:
          type: number
        netChange:
          type: number
        period:
          type: string

    # Services
...
    Notification:
      type: object
      required: [id, title, message, type, isRead, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, warning, success, error]
        isRead:
          type: boolean
        createdAt:
          type: string

    # Audit Logs
...